requirements.txt
# CryptoBotTrader — requirements.txt
# Python 3.12+ (Linux). Пин по мажор/минор ради стабильности.

# --- Web/API ---
flask>=3.0,<4.0
flask-socketio>=5.3.6,<6.0
eventlet>=0.36,<0.37
gunicorn>=21.2,<22.0

# --- Data stack ---
numpy>=1.26
pandas>=2.1
scipy>=1.10

# --- Exchanges ---
ccxt>=4.0

# --- Storage/DB ---
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg2-binary>=2.9,<3.0

# --- Config/validation ---
python-dotenv>=1.0,<2.0
pydantic>=2.6,<3.0

# --- Scheduling ---
apscheduler>=3.10,<4.0

# --- Portfolio optimisation ---
PyPortfolioOpt>=1.5,<2.0

# --- Time-series modelling ---
statsmodels>=0.14,<1.0
pmdarima>=2.0,<3.0
